Add await for the list retrieval
There was a race condition for the refresh request.
When token would get outdated, we would do 2 refresh
requests, as we would call lists and entries endpoints
in parallel. So the second refresh would fail and invalidate
session.

Instead, we want to always do only one refresh
request, and for that, we want to wait for the lists
request to finish, before doing entries