Add list loading directly on the view
I was having a bit of a delay for table loading, as I
moved it to entries endpoint, which waits for lists to load.
Instead I should be adding loading spinner the moment I
am trying to get entries.

I decided to move it out of the store module and directly
into the component, as I think it is more fitting to be
there