MangaList correctly destroys Edit dialog contents
ElementUI destroy-on-close was not reliable, so this
commit makes use of it's open and closed events, to
toggle dialog contents.

The moment dialog opens, we immediately render the contents.
But when closing, I want to wait till transition ends and
then destroy contents, which is why I have to use
closed instead of close

I can also remove closeEditModal method, as destroing dialog body,
means I don't need to manually reset the data anymore